Consolidate ActionTable, its AppState table, and AppLifecycle into one file
The three Bootstrap files were tightly coupled — the appState table is
keyed on AppLifecycle notification names — and small enough to live
together. Merge them into ActionTable.swift and drop the separate
ActionTable.AppState.swift and AppLifecycle.swift.
